rename ep_pluginloader -> pluginloader
previous pluginloader becomes old_pluginloader
