+++ aiida/backends/djsite/db/models.py
@@ -197,8 +197,8 @@ class DbNode(m.Model):
         appropriate subclass.
         """
         from aiida.orm.node import Node
-        from aiida.common.pluginloader import from_type_to_pluginclassname
-        from aiida.common.ep_pluginloader import load_plugin
+        from aiida.common.old_pluginloader import from_type_to_pluginclassname
+        from aiida.common.pluginloader import load_plugin
         from aiida.common import aiidalogger
 
         try:
