+++ aiida/backends/sqlalchemy/models/node.py
@@ -28,7 +28,7 @@ from aiida.backends.sqlalchemy.models.base import Base, _QueryProperty, _AiidaQu
 from aiida.backends.sqlalchemy.models.utils import uuid_func
 
 from aiida.common import aiidalogger
-from aiida.common.ep_pluginloader import load_plugin
+from aiida.common.pluginloader import load_plugin
 from aiida.common.exceptions import DbContentError, MissingPluginError
 from aiida.common.datastructures import calc_states, _sorted_datastates, sort_states
 
@@ -168,7 +168,7 @@ class DbNode(Base):
         Return the corresponding aiida instance of class aiida.orm.Node or a
         appropriate subclass.
         """
-        from aiida.common.pluginloader import from_type_to_pluginclassname
+        from aiida.common.old_pluginloader import from_type_to_pluginclassname
         from aiida.orm.node import Node
 
         try:
