+++ aiida/common/pluginloader.py
@@ -1,327 +1,254 @@
 # -*- coding: utf-8 -*-
-###########################################################################
-# Copyright (c), The AiiDA team. All rights reserved.                     #
-# This file is part of the AiiDA code.                                    #
-#                                                                         #
-# The code is hosted on GitHub at https://github.com/aiidateam/aiida_core #
-# For further information on the license, see the LICENSE.txt file        #
-# For further information please visit http://www.aiida.net               #
-###########################################################################
-import importlib
+"""
+Extension and eventually replacement of the aiida.common.old_pluginloader module
 
-import aiida.common
+Allows loading plugins registered as setuptools entry points by separate pip-installed
+packages.
+defines drop-in replacement functionality to use the old filesystem based and the
+new setuptools based plugin systems in parallel.
+"""
+
+try:
+    from reentry import manager as epm
+except ImportError:
+    import pkg_resources as epm
+
 from aiida.common.exceptions import MissingPluginError
 
 
-logger = aiida.common.aiidalogger.getChild('pluginloader')
+_category_mapping = {
+    'calculations': 'aiida.orm.calculation.job',
+    'data': 'aiida.orm.data',
+    'parsers': 'aiida.parsers.plugins',
+    'schedulers': 'aiida.scheduler.plugins',
+    'transports': 'aiida.transport.plugins',
+    'workflows': 'aiida.workflows',
+    'tools.dbexporters': 'aiida.tools.dbexporters',
+    'tools.dbimporters': 'aiida.tools.dbimporters',
+    'tools.dbexporters.tcod_plugins': 'aiida.tools.dbexporters.tcod_plugins'
+}
 
 
-def from_type_to_pluginclassname(typestr):
+_inv_category_mapping = {v: k for k, v in _category_mapping.iteritems()}
+
+
+def _supercls_for_category(category):
+    from aiida.orm.calculation.job import JobCalculation
+    from aiida.orm.data import Data
+    from aiida.parsers import Parser
+    from aiida.scheduler import Scheduler
+    from aiida.transport import Transport
+    from aiida.orm import Workflow
+    from aiida.tools.dbexporters.tcod_plugins import BaseTcodtranslator
+    supercls_mapping = {
+        'calculations': JobCalculation,
+        'data': Data,
+        'parsers': Parser,
+        'schedulers': Scheduler,
+        'transports': Transport,
+        'workflows': Workflow,
+        'tools.dbexporters.tcod_plugins': BaseTcodtranslator
+    }
+    return supercls_mapping.get(category)
+
+_category_suffix_map = {
+    'calculations': 'Calculation',
+    'tools.dbexporters.tcod_plugins': 'Tcodtranslator'
+}
+
+
+def plugin_list(category):
     """
-    Return the string to pass to the load_plugin function, starting from
-    the 'type' field of a Node.
+    Get a list of plugins for the given category.
+
+    Passing `example` for the category will list all plugins registered under the
+    entry point `aiida.example`.
     """
-    # Fix for base class
-    from aiida.common.exceptions import DbContentError
-    if typestr == "":
-        typestr = "node.Node."
-    if not typestr.endswith("."):
-        raise DbContentError("The type name '{}' is not valid!".format(
-            typestr))
-    return typestr[:-1]  # Strip final dot
 
-def get_query_type_string(plugin_type_string):
+    group = 'aiida.{}'.format(category)
+
+    return [ep.name
+            for ep in epm.iter_entry_points(group=group)]
+
+
+def all_plugins(category):
     """
-    Receives a plugin_type_string, an attribute of subclasses of Node.
-    Checks whether it is a valid type_string and manipulates the string
-    to return a string that in a query returns all instances of a class and
-    all instances of subclasses.
+    find old and new plugins
 
-    :param plugin_type_string: The plugin_type_string
-
-    :returns: the query_type_string
+    If both are available for a given name, the old style plugin takes precedence.
     """
-    from aiida.common.exceptions import DbContentError, InputValidationError
-    if not isinstance(plugin_type_string, basestring):
-        raise InputValidationError("You have to pass as argument")
-    # First case, an empty string for Node:
-    if plugin_type_string == '':
-        query_type_string = ''
-    # Anything else should have baseclass.Class., so at least 2 dots
-    # and end with a dot:
-    elif not(plugin_type_string.endswith('.')) or plugin_type_string.count('.') == 1:
-        raise DbContentError(
-                "The type name '{}' is not valid!".format(plugin_type_string)
-            )
-    else:
-        query_type_string = '{}.'.format('.'.join(plugin_type_string.split('.')[:-2]))
-    return query_type_string
+    from aiida.common.old_pluginloader import existing_plugins
+    from aiida.orm.calculation.job import JobCalculation
+    from aiida.parsers import Parser
+    supercls = _supercls_for_category(category)
+    internal = _category_mapping.get(category)
+    suffix = _category_suffix_map.get(category)
+    plugins = existing_plugins(supercls, internal, suffix=suffix)
+    plugins += [i for i in plugin_list(category) if i not in plugins]
+    return plugins
 
 
-def get_class_typestring(type_string):
+def get_plugin(category, name):
     """
-    Given the type string, return three strings: the first one is
-    one of the first-level classes that the Node can be:
-    "node", "calculation", "code", "data".
-    The second string is the one that can be passed to the DataFactory or
-    CalculationFactory (or an empty string for nodes and codes);
-    the third one is the name of the python class that would be loaded.
+    Return an instance of the class registered under the given name and
+    for the specified plugin category.
+
+    :param category: the plugin category to load the plugin from, e.g.
+          'transport'.
+    :param name: the name of the plugin
     """
-    from aiida.common.exceptions import DbContentError
+    group = 'aiida.{}'.format(category)
 
-    if type_string == "":
-        return ("node", "")
-    else:
-        pieces = type_string.split('.')
-        if pieces[-1]:
-            raise DbContentError("The type string does not end with a dot")
-        if len(pieces) < 3:
-            raise DbContentError("Not enough parts in the type string")
-        return pieces[0], ".".join(pieces[1:-2]), pieces[-2]
+    eps = [ep for ep in epm.iter_entry_points(group=group)
+           if ep.name == name]
 
+    if not eps:
+        raise MissingPluginError(
+            "No plugin named '{}' found for '{}'".format(name, category))
 
-def _existing_plugins_with_module(base_class, plugins_module_path,
-                                  pkgname, basename, max_depth, suffix=None):
-    """
-        Recursive function to return the existing plugins within a given module.
+    if len(eps) > 1:
+        raise MissingPluginError(
+            "Multiple plugins found for '{}' in '{}'".format(name, category))
 
-        :param base_class: Identify all subclasses of the base_class
-        :param plugins_module_path: The path to the folder with the plugins
-        :param pkgname: The name of the package in which you want to search
-        :param basename: The basename of the plugin (sub)class. See also documentation
-            of ``find_module``.
-        :param max_depth: Maximum depth (of nested modules) to be used when
-            looking for plugins
-        :param suffix: The suffix that is appended to the basename when looking
-            for the (sub)class name. If not provided (or None), use the base
-            class name.
-        :return: a list of valid strings that can be used using a Factory or with
-            load_plugin.
-    """
-    import pkgutil
-    import os
+    entrypoint = eps[0]
 
-    if max_depth == 0:
-        return []
-    else:
-        retlist = _find_module(base_class, pkgname, basename, suffix)
-
-        prefix = pkgname + '.'
-        for _, name, ismod in pkgutil.walk_packages([plugins_module_path], prefix=prefix):
-            plugin_name = name[len(prefix):]
-            if ismod:
-                retlist += _existing_plugins_with_module(
-                    base_class, os.path.join(plugins_module_path, plugin_name),
-                    name,
-                    "{}.{}".format(basename, plugin_name) if basename else plugin_name,
-                    max_depth - 1, suffix=suffix)
-
-            # This has to be done anyway, for classes in the __init__ file.
-            this_basename = "{}.{}".format(basename, plugin_name) if basename else plugin_name
-
-            retlist += _find_module(base_class, name, this_basename, suffix)
-
-        return list(set(retlist))
-
-
-def _find_module(base_class, pkgname, this_basename, suffix=None):
-    """
-    Given a base class object, looks for its subclasses inside the package
-    with name pkgname (must be importable), and prepends to the class name
-    the string 'this_basename'.
-
-    If the name of the class complies with the syntax
-    AaaBbb
-    where Aaa is the capitalized name of the containing module (aaa), and
-    Bbb is base_class.__name__, then only 'aaa' is returned instead of
-    'aaa.AaaBbb', to have a shorter name that is anyway accepted by the *Factory
-    functions. If suffix is provided, this is used for comparison (the 'Bbb'
-    string) rather than the base class name)
-
-    :param base_class: Identify all subclasses of the base_class
-    :param pkgname: The name of the package in which you want to search
-    :param basename: The basename of the plugin (sub)class. See also documentation
-        of ``find_module``.
-    :param suffix: The suffix that is appended to the basename when looking
-        for the (sub)class name. If not provided (or None), use the base
-        class name.
-    :return: a list of valid strings, acceptable by the *Factory functions.
-       Does not return the class itself.
-    """
-    import inspect
-
-    retlist = []
-
-    # print ' '*(5-max_depth), '>', pkgname
-    #print ' '*(5-max_depth), ' ', this_basename
-
-    pkg = importlib.import_module(pkgname)
-    for k, v in pkg.__dict__.iteritems():
-        if (inspect.isclass(v) and  # A class
-                    v != base_class and  # Not the class itself
-                issubclass(v, base_class) and  # a proper subclass
-                    pkgname == v.__module__):  # We are importing it from its
-            # module: avoid to import it
-            # from another module, if it
-            # was simply imported there
-            # Try to return the shorter name if the subclass name
-            # has the correct pattern, as expected by the Factory
-            # functions
-            if suffix is None:
-                actual_suffix = base_class.__name__
-            else:
-                actual_suffix = suffix
-
-            if k == "{}{}".format(
-                    pkgname.rpartition('.')[2].capitalize(),
-                    actual_suffix):
-                retlist.append(this_basename)
-            else:
-                retlist.append(
-                    ("{}.{}".format(this_basename, k) if this_basename
-                     else k))
-                #print ' '*(5-max_depth), ' ->', "{}.{}".format(this_basename, k)
-    return retlist
-
-
-def existing_plugins(base_class, plugins_module_name, max_depth=5, suffix=None):
-    """
-    Return a list of strings of valid plugins.
-
-
-    :param base_class: Identify all subclasses of the base_class
-    :param plugins_module_name: a string with the full module name separated
-        with dots that points to the folder with plugins.
-        It must be importable by python.
-    :param max_depth: Maximum depth (of nested modules) to be used when
-            looking for plugins
-    :param suffix: The suffix that is appended to the basename when looking
-        for the (sub)class name. If not provided (or None), use the base
-        class name.
-    :return: a list of valid strings that can be used using a Factory or with
-        load_plugin.
-    """
     try:
-        pluginmod = importlib.import_module(plugins_module_name)
+        plugin = entrypoint.load()
     except ImportError:
-        raise MissingPluginError("Unable to load the plugin module {}".format(
-            plugins_module_name))
+        raise MissingPluginError("Loading the plugin '{}' failed".format(name))
 
-    return _existing_plugins_with_module(base_class,
-                                         pluginmod.__path__[0],
-                                         plugins_module_name,
-                                         "",
-                                         max_depth, suffix)
+    return plugin
 
 
 def load_plugin(base_class, plugins_module, plugin_type):
     """
-    Load a specific plugin for the given base class.
+    load file or extension point plugins using the file plugin interface.
 
-    This is general and works for any plugin used in AiiDA.
+    Prefer file plugins and if not found, translate the arguments to extension point notation
 
-    NOTE: actually, now plugins_module and plugin_type are joined with a dot,
-        and the plugin is retrieved splitting using the last dot of the resulting
-        string.
-    TODO: understand if it is probably better to join the two parameters above
-        to a single one.
-
-    Args:
-        base_class
-            the abstract base class of the plugin.
-        plugins_module
-            a string with the full module name separated with dots
-            that points to the folder with plugins. It must be importable by python.
-        plugin_type
-            the name of the plugin.
-
-    Return:
-        the class of the required plugin.
-
-    Raise:
-        MissingPluginError if the plugin cannot be loaded
-
-    Example:
-       plugin_class = load_plugin(
-           aiida.transport.Transport,'aiida.transport.plugins','ssh.SshTransport')
-
-       and plugin_class will be the class 'aiida.transport.plugins.ssh.SshTransport'
+    :params: Look at the docstring of aiida.common.old_pluginloader.load_plugin for more Info
+    :return: The plugin class
     """
-
-    module_name = ".".join([plugins_module, plugin_type])
-    real_plugin_module, plugin_name = module_name.rsplit('.', 1)
-
-
+    from aiida.common.old_pluginloader import load_plugin as load_old
     try:
-        pluginmod = importlib.import_module(real_plugin_module)
-    except ImportError:
-        raise MissingPluginError("Unable to load the plugin module {}".format(
-            real_plugin_module))
-
-    try:
-        pluginclass = pluginmod.__dict__[plugin_name]
-    except KeyError:
-        raise MissingPluginError("Unable to load the class {} within {}".format(
-            plugin_name, real_plugin_module))
-
-    try:
-        if issubclass(pluginclass, base_class):
-            return pluginclass
-        else:
-            # Quick way of going into the except case
-            err_msg = "{} is not a subclass of {}".format(
-                module_name, base_class.__name__)
-            raise MissingPluginError(err_msg)
-    except TypeError:
-        # This happens when we pass a non-class to issubclass;
-        err_msg = "{} is not a class".format(
-            module_name)
-        raise MissingPluginError(err_msg)
+        plugin = load_old(base_class, plugins_module, plugin_type)
+    except MissingPluginError as e:
+        full_name = plugins_module + '.' + plugin_type
+        catlist = [(c, pm) for c, pm in _category_mapping.iteritems() if pm in full_name]
+        if not catlist:  # find category for new style type string
+            catlist = [(c, 'aiida.orm.' + c) for c in _category_mapping.iterkeys() if c in full_name]
+        if not catlist:
+            raise e
+        category, path = catlist.pop(0)
+        classname = full_name.replace(path, '').strip('.')
+        name = '.'.join(classname.split('.')[:-1])
+        plugin = get_plugin(category, name)
+    return plugin
 
 
 def BaseFactory(module, base_class, base_modname, suffix=None):
     """
-    Return a given subclass of Calculation, loading the correct plugin.
+    Return a plugin class, also find external plugins
 
-    :example: If `module='quantumespresso.pw'`, `base_class=JobCalculation`,
-      `base_modname = 'aiida.orm.calculation.job'`, and `suffix='Calculation'`,
-      the code will first look for a pw subclass of JobCalculation
-      inside the quantumespresso module. Lacking such a class, it will try to look
-      for a 'PwCalculation' inside the quantumespresso.pw module.
-      In the latter case, the plugin class must have a specific name and be
-      located in a specific file:
-      if for instance plugin_name == 'ssh' and base_class.__name__ == 'Transport',
-      then there must be a class named 'SshTransport' which is a subclass of base_class
-      in a file 'ssh.py' in the plugins_module folder.
-      To create the class name to look for, the code will attach the string
-      passed in the base_modname (after the last dot) and the suffix parameter,
-      if passed, with the proper CamelCase capitalization. If suffix is not
-      passed, the default suffix that is used is the base_class class name.
 
-    :param module: a string with the module of the plugin to load, e.g.
-          'quantumespresso.pw'.
-    :param base_class: a base class from which the returned class should inherit.
-           e.g.: JobCalculation
-    :param base_modname: a basic module name, under which the module should be
-            found. E.g., 'aiida.orm.calculation.job'.
-    :param suffix: If specified, the suffix that the class name will have.
-      By default, use the name of the base_class.
+    This is a front end to aiida.common.old_pluginloader.BaseFactory
+    with a fallback to aiida.common.pluginloader.get_plugin
+    check their relative docs for more info.
+
+    Note not all possible notations work for new plugins. Example::
+
+        BaseFactory('quantumespresso.cp', JobCalculation,
+                    'aiida.orm.calculation.job',
+                    suffix='Calculation') # <- finds cp also if new style
+
+        BaseFactory('job.quantumespresso.cp', Calculation,
+                    'aiida.orm.calculation') <- will find cp only if old style
     """
+    from aiida.common.old_pluginloader import BaseFactory as old_factory
     try:
-        return load_plugin(base_class, base_modname, module)
-    except MissingPluginError as e1:
-        # Automatically add subclass name and try again
-        if suffix is None:
-            actual_suffix = base_class.__name__
-        else:
-            actual_suffix = suffix
-        mname = module.rpartition('.')[2].capitalize() + actual_suffix
-        new_module = module + '.' + mname
-        try:
-            return load_plugin(base_class, base_modname, new_module)
-        except MissingPluginError as e2:
-            err_msg = ("Neither {} or {} could be loaded from {}. "
-                       "Error messages were: '{}', '{}'").format(
-                module, new_module, base_modname, e1, e2)
-            raise MissingPluginError(err_msg)
+        return old_factory(module, base_class, base_modname, suffix)
+    except MissingPluginError as e:
+        category = _inv_category_mapping.get(base_modname)
+        if not category:
+            raise e
+        return get_plugin(category, module)
 
+
+def existing_plugins(base_class, plugins_module_name, max_depth=5, suffix=None):
+    """
+    Return a list of plugin names, old and new style
+
+    Refer to aiida.common.old_pluginloader.existing_plugins for more info
+    on behaviour for old style plugins.
+
+    If no old style plugins are found and the plugins_module_name is mappable to a
+    group of entry points, aiida.common.pluginloader.plugin_list is returned
+    """
+    from aiida.common.old_pluginloader import existing_plugins as old_existing
+    plugins = []
+    try:    # try and add internal plugins to the list
+        plugins += old_existing(base_class, plugins_module_name, max_depth=max_depth, suffix=suffix)
+    except MissingPluginError as e:
+        # if there's no entry point for this type of plugin, reraise
+        category = _inv_category_mapping.get(plugins_module_name)
+        if not category:
+            raise e
+        else:
+            pass
+        plugins += plugin_list(category)
+    return plugins
+
+
+def get_class_to_entry_point_map(short_group_name=False):
+    """
+    create a mapping of modules to entry point groups / names
+
+    :param short_group_name: bool, if True the leading 'aiida.' is cut off group names
+    :return: dictionary, keys are modules, values are (group, entrypoint-name)
+    """
+    groups = (g for g in epm.get_entry_map('aiida-core').iterkeys() if g.startswith('aiida'))
+    class_ep_map = {}
+    for group in groups:
+        for ep in epm.iter_entry_points(group):
+            for classname in ep.attrs:
+                key = '.'.join([ep.module_name, classname])
+                if short_group_name:
+                    groupname = '.'.join(group.split('.')[1:])
+                else:
+                    groupname = group
+                value = (groupname, ep.name)
+                class_ep_map[key] = value
+    # ~ module_ep_map = {'.'.join([ep.module_name, ep.attrs]): (group, ep.name) for group in groups for ep in iter_entry_points(group)}
+    return class_ep_map
+
+
+def entry_point_tpstr_from(plugin_class):
+    """
+    gives group and entry point name for a given module if module is registered
+
+    :return: tuple (group, entrypoint-name) if one entry point is found
+    """
+    if isinstance(plugin_class, (str, unicode)):
+        class_path = plugin_class
+        class_name = plugin_class.split('.')[-1]
+    else:
+        class_name = plugin_class.__name__
+        class_path = '.'.join([plugin_class.__module__, class_name])
+    mapping = get_class_to_entry_point_map(short_group_name=True).get(class_path)
+    typestr = None
+    if mapping:
+        group, epname = mapping
+        typestr = '.'.join([_category_mapping[group].replace('aiida.orm.', ''), epname, class_name]) + '.'
+    return typestr
+
+
+def entry_point_from_tpstr(typestring):
+    if typestring.startswith('calculation.job.'):
+        typestring = typestring.split('.', 2)[-1]
+    elif typestring.startswith('calculation.'):
+        typestring = typestring.split('.', 1)[-1]
+    elif typestring.startswith('data.'):
+        typestring = typestring.split('.', 1)[-1]
+    else:
+        raise ValueError('weird typestring')
+    return typestring.split('.', 1)[0]
