+++ aiida/orm/implementation/django/calculation/job/__init__.py
@@ -19,7 +19,7 @@ from aiida.orm.group import Group
 from aiida.orm.implementation.django.calculation import Calculation
 from aiida.orm.implementation.general.calculation.job import (
     AbstractJobCalculation)
-from aiida.common.pluginloader import from_type_to_pluginclassname
+from aiida.common.old_pluginloader import from_type_to_pluginclassname
 from aiida.utils import timezone
