+++ aiida/orm/node.py
@@ -8,7 +8,7 @@
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
 from aiida.orm.implementation import Node
-from aiida.common.pluginloader import from_type_to_pluginclassname
+from aiida.common.old_pluginloader import from_type_to_pluginclassname
 from aiida.orm.implementation.general.node import AttributeManager
