+++ aiida/tools/dbimporters/__init__.py
@@ -14,7 +14,7 @@ def DbImporterFactory(pluginname):
     """
     This function loads the correct DbImporter plugin class
     """
-    from aiida.common.ep_pluginloader import BaseFactory
+    from aiida.common.pluginloader import BaseFactory
     from aiida.tools.dbimporters.baseclasses import DbImporter
 
     return BaseFactory(pluginname, DbImporter, "aiida.tools.dbimporters.plugins")
